fix(api,test): correct HTTP status codes 400→422, Diesel NotFound handling, and test cleanup
- src/api/http/handler/assignment.rs: Change OpenAPI response code 400→422 for validation errors
- src/api/http/handler/assignment_invitation.rs: Change OpenAPI response code 400→422
- src/api/http/handler/auth.rs: Change OpenAPI response code 400→422
- src/api/http/handler/chapter.rs: Change OpenAPI response code 400→422
- src/api/http/handler/chapter_port.rs: Change OpenAPI response code 400→422
- src/api/http/handler/member.rs: Change OpenAPI response codes 400→422 (two endpoints)
- src/api/http/handler/page.rs: Change OpenAPI response code 400→422
- src/api/http/handler/unit.rs: Change OpenAPI response code 400→422
- src/part_impl/rdb_core/result.rs: Convert Diesel NotFound from Unrecoverable to Expected
  - Log warning at call site instead of panicking
  - Return Expected with Args variant and i18n key "error-not-found"
- poprako-util/locales/en-US/main.ftl: Add new error locale keys (not-found, comic-not-found, avatar-version-mismatch, cover-version-mismatch, chapter-pages-already-reserved)
- poprako-util/locales/zh-CN/main.ftl: Add Chinese translations for the new error keys
- src/usecase/page.rs: Use specific error key "error-chapter-pages-already-reserved" instead of generic "error-invalid-page-count"
- src/complex/workset.rs: Change default list limit from u64::MAX to i32::MAX as u64
- src/part_impl/effect_async/chapter.rs: Change default list limit from u64::MAX to i32::MAX as u64
- .gitignore: Add /tests/integration-tests/node_modules and .obsidian/ entries
- tests/AGENTS.md: Document integration-tests directory and test case doc sync requirement
- tests/integration-tests/pnpm-workspace.yaml: Add packages config for standalone workspace
- tests/integration-tests/src/config/env.ts: Fix path resolution for standalone integration-tests location
- tests/integration-tests/src/db/seed.ts: Add cleanupLeftoverRows helper and improved assertion error messages
- tests/integration-tests/src/main.ts: Add cleanup logic before database assertion
- tests/integration-tests/src/state/context.ts: Add commentId and announcementId tracking
- tests/integration-tests/src/suites/allApiSmoke.ts: Replace expectNonServerError with typed expectError/expectSuccessData/expectStatus assertions
- tests/integration-tests/src/suites/projectFlow.ts: Track commentId in test context
- tests/integration-tests/TESTCASES.md: Add integration test case documentation

+++ src/complex/workset.rs
@@ -52,7 +52,7 @@ impl WorksetComplex {
             kind: ComicListKind::All,
             incl_opt: Vec::new(),
             offset: 0,
-            limit: u64::MAX,
+            limit: i32::MAX as u64,
         };
 
         let comic_infos = repo
